+++ lib/openmaize/plugs/authorize.ex
@@ -1,18 +1,10 @@
 defmodule Openmaize.Authorize do
   @moduledoc """
-  Plug to verify that users are authorized to access the requested pages
-  / resources.
+  Plug that performs a basic check that users are authorized to access the
+  requested pages / resources.
 
   Authorization is based on user roles, and so you will need a `role` entry
   in your user model.
-
-  This plug can be used as a first stage in authorizing users, and so you
-  can call further plugs afterwards to make more fine-grained checks. To
-  help these further checks, if authorization is successful, two variables,
-  `path` and `match` are stored in the conn.private.openmaize_vars map.
-  `path` is the full path of the connection and `match` refers to a matching
-  path in the Config.protected map. If no `match` is found, it means that
-  the page is unprotected, and extra Openmaize checks are skipped.
 
   There is one option:
 
@@ -45,10 +37,10 @@ defmodule Openmaize.Authorize do
       conn
     else
       opts = {Keyword.get(opts, :redirects, true)}
-      run(conn, opts, Map.get(assigns, :current_user))
+      run_check(conn, opts, Map.get(assigns, :current_user))
     end
   end
-  defp run(conn, opts, data) do
+  defp run_check(conn, opts, data) do
     get_match(conn) |> permitted?(data) |> authorized?(conn, opts)
   end